+++ inst/tinytest/linear_algebra/test-sd_lc.R
@@ -7,43 +7,42 @@ errorfun <- function(tt) {
 }
 
 # basic tests ====
+coercefun <- list(
+  as.double,
+  as.integer,
+  as.logical
+)
+
 for(j in c(NaN, NA, 0, -1e6, 1e6)) {
-  
-  nobs <- 1e3
-  nvars <- 200
-  n <- nobs * nvars
-  X <- matrix(rnorm(1000), nobs, nvars)
-  vc <- matrix(rnorm(1000), nvars, nvars)
-  
-  out <- numeric(nobs)
-  for(i in 1:nobs) {
-    a <- matrix(X[i,, drop = FALSE], ncol = 1)
-    out[i] <- sqrt(t(a) %*% vc %*% a)
+  for(k in seq_along(coercefun)) {
+    
+    nobs <- 20
+    nvars <- 10
+    n <- nobs * nvars
+    X <- matrix(coercefun[[k]](rnorm(100)), nobs, nvars)
+    vc <- cov(as.data.frame(X))
+    
+    out <- rowSums((X %*% vc) * X) |> sqrt()
+    out[is.na(out)] <- j
+    expect_equal(
+      out,
+      sd_lc(X, vc, j)
+    ) |> errorfun()
+    enumerate <- enumerate + 1L
+    
   }
-  out[is.na(out)] <- j
-  expect_equal(
-    out,
-    sd_lc(X, vc, j)
-  ) |> errorfun()
-  enumerate <- enumerate + 1L
   
 }
 
-
-
 vc <- datasets::ability.cov$cov
-nobs <- 500
+nobs <- 100
 nvars <- nrow(vc)
 n <- nobs * nvars
 
-X <- matrix(rnorm(1000), nobs, nvars)
+X <- matrix(rnorm(100), nobs, nvars)
 
 
-out <- numeric(nobs)
-for(i in 1:nobs) {
-  a <- matrix(X[i,, drop = FALSE], ncol = 1)
-  out[i] <- sqrt(t(a) %*% vc %*% a)
-}
+out <- rowSums((X %*% vc) * X) |> sqrt()
 expect_equal(
   out,
   sd_lc(X, vc)
@@ -51,10 +50,54 @@ expect_equal(
 enumerate <- enumerate + 1L
 
 
+# missing/special values tests ====
+
+for(i in 1:10) { # multiple iterations to take into account randomness
+  
+  # make data:
+  vc <- datasets::ability.cov$cov
+  nobs <- 100
+  nvars <- nrow(vc)
+  n <- nobs * nvars
+  
+  
+  # test real:
+  X <- matrix(sample(c(1:6, NaN, NA, -Inf, Inf)), nobs, nvars)
+  out <- rowSums((X %*% vc) * X) |> sqrt()
+  expect_equal(
+    out,
+    sd_lc(X, vc)
+  ) |> errorfun()
+  
+  
+  # test integer:
+  X <- matrix(sample(c(1:9, NA_integer_)), nobs, nvars)
+  out <- rowSums((X %*% vc) * X) |> sqrt()
+  expect_equal(
+    out,
+    sd_lc(X, vc)
+  ) |> errorfun()
+  
+  
+  # test logical:
+  X <- matrix(sample(c(TRUE, FALSE, NA), 10, TRUE), nobs, nvars)
+  out <- rowSums((X %*% vc) * X) |> sqrt()
+  expect_equal(
+    out,
+    sd_lc(X, vc)
+  ) |> errorfun()
+  
+  
+  enumerate <- enumerate + 3L
+  
+}
+
+
 # distributional tests ====
 funlist <- list( # normal distribution already tested above, so not needed here
   \(p) rbeta(1000, p, p+1),
-  \(p) rbinom(1000, 10, p/10),
+  \(p) rbinom(1000, 1, p/10) |> as.logical(), # Bernoulli
+  \(p) rbinom(1000, 10, p/10) |> as.integer(),
   \(p) rcauchy(1000, 0, p),
   \(p) rchisq(1000, p),
   \(p) rexp(1000, p),
@@ -63,7 +106,8 @@ funlist <- list( # normal distribution already tested above, so not needed here
   \(p) rgeom(1000, p/10),
   \(p) rhyper(1000, 100-p, (100 - p) - p*2, ((100 - p) - p*2) - p*3),
   \(p) rlnorm(1000, 1, p),
-  \(p) rpois(1000, p),
+  \(p) rnbinom(1000, p, 1/p) |> as.integer(),
+  \(p) rpois(1000, p) |> as.integer(),
   \(p) rt(1000, p),
   \(p) runif(1000, p, p + 1),
   \(p) rweibull(1000, p)
@@ -122,7 +166,7 @@ Xlist <- list(
 for(i in Xlist) {
   expect_error(
     sd_lc(i, vc),
-    pattern = "`X` must be a numeric matrix"
+    pattern = "`X` must be a numeric or logical matrix"
   ) |> errorfun()
   enumerate <- enumerate + 1L
 }