+++ src/C_sd_lc.c
@@ -3,43 +3,108 @@
 #include <R_ext/Error.h>
 
 SEXP C_sd_lc(
-  SEXP w, SEXP vc, SEXP nvars0, SEXP nobs0, SEXP bad_rp0
+  SEXP w, SEXP vcdiag, SEXP vc, SEXP nobs0, SEXP nvars0, SEXP bad_rp0
 ) {
   
-  // For the mathematical formula this code is based on,
-  // See page 76 of:
-  // Applied Multivariate Statistical Analysis (6th Edition), by Richard Johnson & Dean Wichern (2001)
+  // note: x[a, b] == x[a + b * nobs]
   
   int nvars = INTEGER(nvars0)[0];
   int nobs = INTEGER(nobs0)[0];
   double bad_rp = REAL(bad_rp0)[0];
-  
-  const double *pw = REAL_RO(w);
   const double *pvc = REAL_RO(vc);
-  double sum1, sum2;
+  const double *pvcdiag = REAL_RO(vcdiag);
+  double var, covar;
+  R_xlen_t index, i_t_nvars;
   
   SEXP out = PROTECT(Rf_allocVector(REALSXP, nobs));
   double *pout = REAL(out);
-
-  for(int k = 0; k < nobs; ++k) {
-    sum2 = 0.0;
-    for(int i = 0; i < nvars; ++i) {
-      sum1 = 0.0;
-      for(int j = 0; j < nvars; ++j) {
-        sum1 += pw[k + j * nobs] * pvc[j + i * nvars];
+  
+  if(TYPEOF(w) == REALSXP) { // start decimal function
+    const double *pw = REAL_RO(w);
+    double weight;
+    for(int k = 0; k < nobs; ++k) {
+      
+      // variance:
+      var = 0.0;
+      for(int i = 0; i < nvars; ++i) {
+        weight = pw[k + i * nobs];
+        var += pvcdiag[i] * weight * weight;
       }
-      sum2 += sum1 * pw[k + i * nobs];
+      
+      // covariance:
+      covar = 0;
+      for(int i = 0; i < (nvars - 1); ++i) {
+        index = k + i * nobs;
+        i_t_nvars = i * nvars;
+        for(int j = (i + 1); j < nvars; ++j) {
+          covar += pw[k + j * nobs] * pw[index] * pvc[j + i_t_nvars];
+        }
+      }
+      
+      // combine:
+      var = var + 2 * covar;
+      if(var < 0) {
+        pout[k] = bad_rp;
+      }
+      else {
+        pout[k] = sqrt(var);
+      }
+      
     }
-    if(sum2 < 0) {
-      pout[k] = bad_rp;
+  } // end decimal function
+  
+  if(TYPEOF(w) == INTSXP || TYPEOF(w) == LGLSXP) { // start integer function
+    int isNA;
+    const int *pw = INTEGER_RO(w);
+    int weight;
+    
+    for(int k = 0; k < nobs; ++k) {
+      
+      // NAs:
+      isNA = 0;
+      for(int i = 0; i < nvars; ++i) {
+        if(pw[k + i * nobs] == NA_INTEGER) {
+          isNA = 1;
+          break;
+        }
+      }
+      if(isNA) {
+        // return NaN
+        pout[k] = NA_REAL;
+      }
+      else { // start regular integer computation
+        // variance:
+        var = 0.0;
+        for(int i = 0; i < nvars; ++i) {
+          weight = pw[k + i * nobs];
+          var += pvcdiag[i] * weight * weight;
+        }
+        
+        // covariance:
+        covar = 0;
+        for(int i = 0; i < (nvars - 1); ++i) {
+          index = k + i * nobs;
+          i_t_nvars = i * nvars;
+          for(int j = (i + 1); j < nvars; ++j) {
+            covar += pw[k + j * nobs] * pw[index] * pvc[j + i_t_nvars];
+          }
+        }
+        
+        // combine:
+        var = var + 2 * covar;
+        if(var < 0) {
+          pout[k] = bad_rp;
+        }
+        else {
+          pout[k] = sqrt(var);
+        }
+      } // end regular integer computation
     }
-    else {
-      pout[k] = sqrt(sum2);
-    }
-  }
+  } // end integer function
+  
+  
   
   UNPROTECT(1);
   return out;
   
 }
-